transformers restructuee
